+++ thebleep/rules/yarn_alias.py
@@ -1,5 +1,5 @@
 import re
-from thebleep.utils import replace_argument, for_app
+from thebleep.utils import quote_words, replace_argument, for_app
 
 
 @for_app('yarn', at_least=1)
@@ -11,4 +11,4 @@ def get_new_command(command):
     broken = command.script_parts[1]
     fix = re.findall(r'Did you mean [`"](?:yarn )?([^`"]*)[`"]', command.output)[0]
 
-    return replace_argument(command.script, broken, fix)
+    return replace_argument(command.script, broken, quote_words(fix))